% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{alter_pathway_elements}
\alias{alter_pathway_elements}
\title{alter_pathway_elements}
\usage{
alter_pathway_elements(homology.table, human.pathway)
}
\arguments{
\item{homology.table}{Homology table for the pathway molecules of the two species - species1 and species2}

\item{human.pathway}{A list with human pathway path data where each sublist denotes a path of the pathway. This is used as background data.}
}
\value{
This function returns the pathway data by altering the pathway molecules by the homology molecules.
}
\description{
This function alters human pathway elements by homology molecules of other species.
}
\details{
This function alters human pathway elements by homology molecules of other species.
}
\examples{
P<-list(c('P','G','A','N'), c('P','C','X','T'), c('P','M','N','D'))
Q<-list(c('Q','M','O','T','L'), c('Q','U','T','F'), c('Q','M','N','I','B'))
R<-list(c('R','X','B','K'), c('R','G','H','O'), c('R','U','C','E'))
QQ<-list(c('Q','M','O','T','L'), c('Q','U','T','F'), c('Q','M','N','I','B'))
S<-list(c('S','M','NN'), c('S','N','O','B1'))
U<-list(c('SS','OO','nn'), c('ss','oo','hh'))
V<-list(c('V','X','B','D','J','E','K'), c('V','M','O','F','D','G'), c('V','B','C'))
human.pathway<-list("P"=P, "Q"=Q, "R"=R, "QQ"=QQ, "S"=S, "U"=U, "V"=V)
homology.table<-data.frame('human_mol'=c('B1','C','X','R',LETTERS), 'other_mol'=c('b','c1','x1','r1',letters))
homology.pathway<-alter_pathway_elements(homology.table, human.pathway)
homology.pathway
}
