% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{convert_geneID_to_geneSymbol}
\alias{convert_geneID_to_geneSymbol}
\title{convert_geneID_to_geneSymbol}
\usage{
convert_geneID_to_geneSymbol(
  user.cell.data,
  species = "hsapiens",
  geneSymbol.toMapping = "external_gene_name",
  geneID.forMapping = "ensembl_gene_id",
  collapse.method = "max"
)
}
\arguments{
\item{user.cell.data}{Cell/tissue profile with valid gene IDs}

\item{species}{The species of the given gene IDs or cell/tissue profile, default is "hsapiens"}

\item{geneSymbol.toMapping}{The valid gene ID names (that has ensembl id mapping in the biomaRt database) to which the user.cell.data will be converted, default is "external_gene_name". For more details to get the valid gene ids for a species please see the find_valid_geneID() function of the package.}

\item{geneID.forMapping}{The valid gene ID names (that has ensembl id mapping in the biomaRt database) of the user.cell.data, default is "ensembl_gene_id". For more details to get the valid gene ids for a species please see the find_valid_geneID() function of the package.}

\item{collapse.method}{Used when one external_gene_name has more then one probes, usually two options are used, maximum probe value (i.e., "max") or average of probe values (i.e., "mean"), default is "max".}
}
\value{
This function returns cell/tissue profiles with converted gene symbols
}
\description{
This function converts either a vector of gene IDs, a matrix of gene or a named cell / tissue expression profile or matrix to ensembl IDs
}
\details{
This function converts gene IDs of a named cell/tissue expression profile to gene symbols
}
\examples{
query.data<-matrix(sample(1:10, 100, replace=TRUE),10,10)
rownames(query.data)<-c("ENSG00000160202", "ENSG00000109846", "ENSG00000100053", "ENSG00000007372", "ENSG00000181449", "ENSG00000117707", "ENSG00000138083", "ENSG00000150938", "ENSG00000244752", "ENSG00000101144")
colnames(query.data)<-c("cell1", "cell1", "cell1", "cell2", "cell2", "cell2", "cell3", "cell3", "cell3", "cell3")
convert_geneID_to_geneSymbol(query.data)

}
