% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{display_pathway_activity_score}
\alias{display_pathway_activity_score}
\title{display_pathway_activity_score}
\usage{
display_pathway_activity_score(
  pathway.activity.score,
  homology.data,
  highly.expressed.rp = rp.median.exp.4,
  rp.expression.range = c(5, 15)
)
}
\arguments{
\item{pathway.activity.score}{The ranking metric result returned by 'get_pathway_ranking_metric' function.}

\item{homology.data}{A list containing the homology data (i.e., homology table, homology matrix and species2 name) for the two species.}

\item{highly.expressed.rp}{A vector containing a list of receptor proteins that are generally highly expressed in many cells and/or tissues. Default is rp.median.exp.4.}

\item{rp.expression.range}{A vector containing an expression range (in the format of c(min, max)) of the receptor proteins to show at the y axis.}
}
\description{
This function plots the pathway activity score values for each cell type or tissue in a 2D plane. This function uses generally highly expressed receptor data from the package to separate the cell/tissue specific active pathways and generally active pathways in many cells. In the figure, the black and gray colours represent cell/tissue specifically active pathways and generally active pathways in many cells respectively.
}
\details{
This function plots the pathway activity score values for each cell type or tissue in a 2D plane. This function uses generally highly expressed receptor data from the package to separate the cell/tissue specific active pathways and generally active pathways in many cells. In the figure, the black and gray colours represent cell/tissue specifically active pathways and generally active pathways in many cells respectively.
}
\examples{
#Pre-process the 'tooth.epi.E13.5' data
tooth.epi.E13.5.processed.data<-preprocess_querydata_new(cell.tissue.data = tooth.epi.E13.5, exp.cutoff.th = 5.0, species="mmusculus")
#Generate the homology data for the two species - 'hsapiens' and 'mmusculus'
mouse.homology.data<-generate_homology_data(species1 = "hsapiens", species2 = "mmusculus")
#Generate the mouse homology pathway path data
mouse.homology.pathway.path<-generate_homology_pathways(species.homology.data = mouse.homology.data, pathway.path = pathway.path.new)
#Identify active pathway paths of the processed query data
tooth.epi.E13.5.active.pathway<-identify_active_pathway_path_new(pathway.path = mouse.homology.pathway.path, processed.query.data = tooth.epi.E13.5.processed.data)
#Get pathway activity score (i.e., pathway name gene expression and pathway specific gene count proportion) of the processed query data
tooth.epi.E13.5.active.pathway.score<-get_pathway_activity_score_new(active.pathway.path = tooth.epi.E13.5.active.pathway, processed.query.data = tooth.epi.E13.5.processed.data, homology.data = mouse.homology.data)
#Plot the activity score result (i.e., pathway name gene expression versus pathway specific gene count proportion) in a 2D plane (black=specifically active, gray=generally active)
display_pathway_activity_score(pathway.activity.score = tooth.epi.E13.5.active.pathway.score, homology.data = mouse.homology.data)
#To separate the top ranked pathways we can do this
abline(v=0.2, h=10, lty=2, col="black")

}
