% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{draw_active_pathways}
\alias{draw_active_pathways}
\title{draw_active_pathways}
\usage{
draw_active_pathways(active.pathway.ppi.data.frame)
}
\arguments{
\item{active.pathway.ppi.data.frame}{A list of sublist containing active pathway data frame for each pathway of each cell / tissue.}
}
\value{
This function draws all the pathways as tree layout and stores in one pdf file for each cell / tissue and finally returns null.
}
\description{
This function draws all the pathways as tree layout and stores in one pdf file for each cell / tissue.
}
\details{
This function draws all the pathways as tree layout and stores in one pdf file for each cell / tissue.
}
\examples{
#Pre-process the 'tooth.epi.E13.5' data
tooth.epi.E13.5.processed.data<-preprocess_querydata_new(cell.tissue.data = tooth.epi.E13.5, exp.cutoff.th = 5.0, species="mmusculus")
#Generate the mouse homology pathway path data
mouse.homology.pathway.path<-generate_homology_pathways(species1 = "hsapiens", species2 = "mmusculus", pathway.path = pathway.path.new)
#Identify active pathway paths of the processed query data
tooth.epi.E13.5.active.pathway<-identify_active_pathway_path(pathway.path = mouse.homology.pathway.path, processed.query.data = tooth.epi.E13.5.processed.data)
#Generate the active pathway paths data frame
tooth.epi.E13.5.active.pathway.df<-generate_pathway_ppi_data_frame(active.pathway.path = tooth.epi.E13.5.active.pathway)
#Now draw the pathways as tree layout in one pdf file
draw_active_pathways(tooth.epi.E13.5.active.pathway.df)
}
