% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{generate_homology_gene_set}
\alias{generate_homology_gene_set}
\title{generate_homology_gene_set}
\usage{
generate_homology_gene_set(species.homology.matix, species1.genes)
}
\arguments{
\item{species.homology.matix}{A sparse matrix containing the homology genes information for the two species.}

\item{species1.genes}{A vector containing set of genes for species1.}
}
\value{
This function returns a vector of homology gene set for species2 with respect to the list of genes of species1.
}
\description{
This function generates the homology gene set for species2 with respect to the list of genes of species1.
}
\details{
This function generates the homology gene set for species2 with respect to the list of genes of species1.
}
\examples{
human.zebrafish.homology.data<-generate_homology_data(species1 = "hsapiens", species2 = "drerio")
human.genes<-c("PAX6", "SOX2", "PROX1", "FGFR1", "BMP7", "MAPK1", "ACE2")
zebrafish.selected.homology.genes<-generate_homology_gene_set(species.homology.matix = human.zebrafish.homology.data[[2]], species1.genes = human.genes)
zebrafish.selected.homology.genes
}
