% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{generate_homology_pathways}
\alias{generate_homology_pathways}
\title{generate_homology_pathways}
\usage{
generate_homology_pathways(species1 = "hsapiens", species2, pathway.path)
}
\arguments{
\item{species1}{Species 1 name, in the form 'hsapiens'. Default is ''hsapiens'.}

\item{species2}{Species 2 name, in the form 'hsapiens'.}

\item{pathway.path}{A list with human pathway path data where each sublist denotes a path of the pathway. This is used as a background data.}
}
\value{
This function returns the homology pathways by altering the pathway molecules by the homology molecules.
}
\description{
This function generates homology pathways by altering the molecules of human pathway paths.
}
\details{
This function generates homology pathways by altering the molecules of human pathway paths.
}
\examples{
drerio.homology.pathway.path<-generate_homology_pathways(species1 = "hsapiens", species2 = "drerio", pathway.path = pathway.path.new)
drerio.homology.pathway.path[[1]]
}
