% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{generate_pathway_molecule_homology_table}
\alias{generate_pathway_molecule_homology_table}
\title{generate_pathway_molecule_homology_table}
\usage{
generate_pathway_molecule_homology_table(
  species1 = "hsapiens",
  species2,
  species1.pathway.molecule
)
}
\arguments{
\item{species1}{Species 1 name, in the form 'hsapiens'}

\item{species2}{Species 2 name, in the form 'hsapiens'}

\item{species1.pathway.molecule}{A vector of pathway molecules for species1}
}
\value{
This function returns a homology table for the list of molecules of species1 of the two species.
}
\description{
This function creates a pathway molecule homology table for the two species with respect to the list of molecules of species1 by retrieving the homology table.
}
\details{
This function creates a pathway molecule homology table for the two species with respect to the list of molecules of species1 by retrieving the homology table.
}
\examples{
human.genes<-c("PAX6", "SOX2", "PROX1", "FGFR1", "BMP7", "MAPK1", "ACE2")
human.zebrafish.selected.genes.homology.table<-generate_pathway_molecule_homology_table(species1 = "hsapiens", species2 = "drerio", species1.pathway.molecule = human.genes)
human.zebrafish.selected.genes.homology.table
}
