% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{generate_pathway_ppi_data_frame}
\alias{generate_pathway_ppi_data_frame}
\title{generate_pathway_ppi_data_frame}
\usage{
generate_pathway_ppi_data_frame(active.pathway.path)
}
\arguments{
\item{active.pathway.path}{A list of sublist containing active pathway path data for each cell / tissue.}
}
\value{
This function returns pathway PPI data frame from the active pathway data to draw pathway figures using cytoscape.
}
\description{
This function generates pathway PPI data frame from the active pathway data to draw pathway figures using cytoscape.
}
\details{
This function generates pathway PPI data frame from the active pathway data to draw pathway figures using cytoscape.
}
\examples{
#Pre-process the 'tooth.epi.E13.5' data
tooth.epi.E13.5.processed.data<-preprocess_querydata_new(cell.tissue.data = tooth.epi.E13.5, exp.cutoff.th = 5.0, species="mmusculus")
#Generate the mouse homology pathway path data
mouse.homology.pathway.path<-generate_homology_pathways(species1 = "hsapiens", species2 = "mmusculus", pathway.path = pathway.path.new)
#Identify active pathway paths of the processed query data
tooth.epi.E13.5.active.pathway<-identify_active_pathway_path_new(pathway.path = mouse.homology.pathway.path, processed.query.data = tooth.epi.E13.5.processed.data)
#Generate the active pathway paths data frame
tooth.epi.E13.5.active.pathway.df<-generate_pathway_ppi_data_frame(active.pathway.path = tooth.epi.E13.5.active.pathway)
tooth.epi.E13.5.active.pathway.df[[1]][1]
}
