% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{get_ENSEMBL_gene_list_new}
\alias{get_ENSEMBL_gene_list_new}
\title{get_ENSEMBL_gene_list_new}
\usage{
get_ENSEMBL_gene_list_new(species)
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}
}
\value{
This helper function returns a data frame with one column containing all Ensembl gene IDs for the given species
}
\description{
This helper function returns a data frame with one column containing all Ensembl gene IDs for the given species
}
\details{
This helper function returns a data frame with one column containing all Ensembl gene IDs for the given species
}
\examples{
human_ensembl_genes <- get_ENSEMBL_gene_list_new('hsapiens')
head(human_ensembl_genes)

}
