% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{get_homology_matrix_new}
\alias{get_homology_matrix_new}
\title{get_homology_matrix_new}
\usage{
get_homology_matrix_new(
  species1,
  species2,
  homology.table,
  seq.identity = FALSE
)
}
\arguments{
\item{species1}{Species 1 name, in the form 'hsapiens'}

\item{species2}{Species 2 name, in the form 'hsapiens'}

\item{homology.table}{Homology table for two species}

\item{seq.identity}{Boolean of whether or not  to retrieve sequence identity information, defaults to FALSE}
}
\value{
This function returns a sparse matrix where rows represent the genes in species 1 and columns the genes in species 2.
}
\description{
This function creates a homology matrix between two species by retrieving the homology table and then converting it into a sparse matrix.
}
\details{
This function creates a homology matrix between two species by retrieving the homology table and then converting it into a sparse matrix.
}
\examples{
human.zebrafish.homology.table<-get_homology_table_new('hsapiens','drerio')
human.zebrafish.homology.matix <- get_homology_matrix_new('hsapiens','drerio', human.zebrafish.homology.table)
print(data.matrix(human.zebrafish.homology.matix[1:5,1:5]))
}
