% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{get_pathway_activity_score_new}
\alias{get_pathway_activity_score_new}
\title{get_pathway_activity_score_new}
\usage{
get_pathway_activity_score_new(
  active.pathway.path,
  processed.query.data,
  homology.data,
  hk.gene = housekeeping.gene
)
}
\arguments{
\item{active.pathway.path}{A list of active pathway path data for each cell/tissue type as returned by the function 'identify_active_pathway_path'.}

\item{processed.query.data}{A list with expressed query data where each sublist corresponds for each cell/tissue type as returned by the function 'preprocess_querydata'.}

\item{homology.data}{A list containing the homology data (i.e., homology table, homology matrix and species2 name) for the two species.}

\item{hk.gene}{A vector consisting of housekeeping genes. Default is housekeeping.gene. This data is loaded automatically with the SPAGI package. This data was generated using the gene expression profiles of different cell types and/or tissues from the ENCODE human and mouse project.}
}
\value{
This function returns a list of pathway activity score for each cell/tissue type.
}
\description{
This function generates pathway activity score (i.e., pathway name gene expression versus pathway expressed specific gene count proportion) of the active pathways for each cell/tissue type. It uses active pathway path and processed query data with the homology information to generate the activity score. It also uses housekeeping gene data from SPAGI package to get the cell/tissue expressed specific pathway gene count proportion.
}
\details{
This function generates pathway activity score (i.e., pathway name gene expression versus pathway expressed specific gene count proportion) of the active pathways for each cell/tissue type. It uses active pathway path and processed query data with the homology information to generate the activity score. It also uses housekeeping gene data from SPAGI package to get the cell/tissue expressed specific pathway gene count proportion.
}
\examples{
#Pre-process the 'tooth.epi.E13.5' data
tooth.epi.E13.5.processed.data<-preprocess_querydata_new(cell.tissue.data = tooth.epi.E13.5, exp.cutoff.th = 5.0, species="mmusculus")
#Generate the homology data for the two species - 'hsapiens' and 'mmusculus'
mouse.homology.data<-generate_homology_data(species1 = "hsapiens", species2 = "mmusculus")
#Generate the mouse homology pathway path data
mouse.homology.pathway.path<-generate_homology_pathways(species.homology.data = mouse.homology.data, pathway.path = pathway.path.new)
#Identify active pathway paths of the processed query data
tooth.epi.E13.5.active.pathway<-identify_active_pathway_path_new(pathway.path = mouse.homology.pathway.path, processed.query.data = tooth.epi.E13.5.processed.data)
#Get pathway activity score (i.e., pathway name gene expression and pathway specific gene count proportion) of the processed query data
tooth.epi.E13.5.active.pathway.score<-get_pathway_activity_score_new(active.pathway.path = tooth.epi.E13.5.active.pathway, processed.query.data = tooth.epi.E13.5.processed.data, homology.data = mouse.homology.data)
head(summary(tooth.epi.E13.5.active.pathway.score$E13.5_Epi))

}
