% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{make_ENSEMBL_symbol_map_new}
\alias{make_ENSEMBL_symbol_map_new}
\title{make_ENSEMBL_symbol_map_new}
\usage{
make_ENSEMBL_symbol_map_new(species, geneSymbol.toMapping, geneID.forMapping)
}
\arguments{
\item{species}{Species name in the form 'hsapiens'}

\item{geneSymbol.toMapping}{The valid gene ID names to which the user cell data will be converted, default is "external_gene_name"}

\item{geneID.forMapping}{The valid gene ID names of the user cell data, default is "ensembl_gene_id"}
}
\value{
This helper function returns a data frame with two columns containing all gene symbols and their corresponding gene IDs for the given species
}
\description{
This helper function returns a data frame with two columns containing all gene symbols and their corresponding gene IDs for the given species
}
\details{
This helper function returns a data frame with two columns containing all gene symbols and their corresponding gene IDs for the given species
}
\examples{
human.ensembl.symbol.map<-make_ENSEMBL_symbol_map_new('hsapiens', geneSymbol.toMapping="external_gene_name", geneID.forMapping="ensembl_gene_id")
head(human.ensembl.symbol.map)

}
