% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathway.path.new}
\alias{pathway.path.new}
\title{Pathway path data}
\format{
An object of list containing pathway paths that is used as background data.
}
\usage{
data(pathway.path.new)
}
\description{
Pathway path used as background data, generated using STRINGdb PPI.
}
\examples{
head(summary(pathway.path.new)
}
\keyword{datasets}
