% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rp.median.exp.4}
\alias{rp.median.exp.4}
\title{Receptor proteins that are generally highly expressed in many cells / tissues
Generated by utilizing human and mouse encode data sets from the "pathway_cleaning.R" additional file
Used as background data}
\format{
An object of matrix containing in normalized and log2 format gene expression value of mouse tooth epithelium cell.
}
\usage{
data(rp.median.exp.4)
}
\description{
Receptor proteins that are generally highly expressed in many cells / tissues
Generated by utilizing human and mouse encode data sets from the "pathway_cleaning.R" additional file
Used as background data
}
\examples{
head(rp.median.exp.4)
}
\references{
The ENCODE Project Consortium, Nature 2012 Sep 6;489(7414):57-74
(\href{https://www.ncbi.nlm.nih.gov/pubmed/22955616}{PubMed})

The Mouse ENCODE Consortium, Genome Biology 2012 Aug 13;13(8):418
(\href{https://pubmed.ncbi.nlm.nih.gov/22889292/}{PubMed})
}
\keyword{datasets}
