% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{show_distribution}
\alias{show_distribution}
\title{show_distribution}
\usage{
show_distribution(
  cell.tissue.data,
  data.format = "matrix",
  experiment.descriptor = NULL
)
}
\arguments{
\item{cell.tissue.data}{An expression matrix with replicated column headers per-replicate. It is assumed that all query data are in RPKM/FPKM/CPM and log normalized form. Also assume that gene ids are official gene symbols. For the matrix, rows denote the genes and the columns denote the cell types or tissues. Duplicate column names are expected in this case denoting replicate samples. All the replicate samples for a specific cell or tissue should have identical column names, otherwise the experiment.descriptor parameter should be used to identify replicate samples of a specific cell type or tissue.}

\item{data.format}{Format of cell.tissue.data. Default is "matrix".}

\item{experiment.descriptor}{A vector corresponding to the matrix column names of cell.tissue.data, containing the cell type or tissues of each sample. The names should be identical for a specific cell or tissue.  Defaults to NULL.}
}
\value{
This function displays the distribution of each cell / tissue based on their replicates' average value. It returns null value.
}
\description{
This function helps to look for the distribution for all cells / tissues based on their replicates' average value.
}
\details{
This function helps to look for the distribution for all cells / tissues based on their replicates' average value.
}
\examples{
show_distribution(cell.tissue.data = tooth.epi.E13.5)
abline(v=5,col="red")
}
