% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tooth.epi.E13.5}
\alias{tooth.epi.E13.5}
\title{Data from mouse embryonic dental epithelium cell at E13.5,
contains transcriptomic profiles microarray data of mouse tooth epithelium cell at embryonic day E13.5.
The cell has 3 biological replicates.}
\format{
An object of matrix containing in normalized and log2 format gene expression value of mouse tooth epithelium cell.
}
\usage{
data(tooth.epi.E13.5)
}
\description{
Data from mouse embryonic dental epithelium cell at E13.5,
contains transcriptomic profiles microarray data of mouse tooth epithelium cell at embryonic day E13.5.
The cell has 3 biological replicates.
}
\examples{
head(tooth.epi.E13.5)
}
\references{
Daniel J O'Connell et al, Science Signaling, 2012 Jan 10;5(206):ra4.
(\href{https://pubmed.ncbi.nlm.nih.gov/22234613/}{PubMed})
}
\keyword{datasets}
