% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\name{generate_homology_pathways}
\alias{generate_homology_pathways}
\title{generate_homology_pathways}
\usage{
generate_homology_pathways(species.homology.data, pathway.path)
}
\arguments{
\item{species.homology.data}{A list containing the homology data (i.e., homology table, homology matrix and species2 name) for the two species.}

\item{pathway.path}{A list with human pathway path data where each sublist denotes a path of the pathway. This is used as a background data.}
}
\value{
This function returns the homology pathways by altering the pathway molecules by the homology molecules.
}
\description{
This function generates homology pathways by altering the molecules of human pathway paths.
}
\details{
This function generates homology pathways by altering the molecules of human pathway paths. The human pathway path data is automatically loaded with the package.
}
\examples{
human.zebrafish.homology.data<-generate_homology_data(species1 = "hsapiens", species2 = "drerio")
zebrafish.homology.pathway.path<-generate_homology_pathways(species.homology.data = human.zebrafish.homology.data, pathway.path = pathway.path.new)
head(zebrafish.homology.pathway.path[[1]])
}
