% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spagi2_master.R
\docType{package}
\name{spagi2-package}
\alias{spagi2-package}
\alias{spagi2}
\alias{SPAGI2}
\title{SPAGI2: Identification of active signalling pathways for cross-species by integrating gene expression and human protein interaction network}
\description{
SPAGI2 is an R package for cross species active signalling pathway identification for a gene expression profile. This package contains the neccessary R code to run SPAGI2 as described in "SPAGI2: Identification of active signalling pathways for cross-species by integrating gene expression and human protein interaction network". SPAGI2 is implemented within a helpful framework to identify active pathways from gene expression profiles.
}
\examples{
## Do a sample analysis using mouse embryonic dental epitheliam cell at E13.5 microarray gene expression data.

## Here we will use "pathway.path.new" as background data from the SPAGI2 repository.
## Also we will use "tooth.epi.E13.5" as query microarray gene expression data. This data is for mouse embryonic dental epitheliam cell at E13.5.
## These data sets are loaded automatically with the package.

## Pre-process the query data (tooth.epi.E13.5), The data has already been made in normalized and log2 scale format. Here, we will use the expression cutoff as 5.0 of the query data. Also we have already made the replicate names same for the data.
tooth.epi.E13.5.processed.data<-preprocess_querydata_new(cell.tissue.data = tooth.epi.E13.5, exp.cutoff.th = 5.0, species="mmusculus")
## Generate the mouse homology pathway path data. Here we will use the background human pathway.path.new data to get the mouse homology pathway path data. It will take little bit more time to access the biomaRt ensembl server.
mouse.homology.pathway.path<-generate_homology_pathways(species1 = "hsapiens", species2 = "mmusculus", pathway.path = pathway.path.new)
## Identify active pathway paths of the processed query data. Here, we will use the background pathway.path.new data to get the active pathway paths of the processed query data.
tooth.epi.E13.5.active.pathway<-identify_active_pathway_path(pathway.path = mouse.homology.pathway.path, processed.query.data = tooth.epi.E13.5.processed.data)
## Get active pathway ranking metric (i.e., activity score and number of downstream transcription factors). Here we will use the tooth.epi.E13.5.active.pathway and tooth.epi.E13.5.processed.data data sets to get the acitve pathway ranking metric. Also we will use a high expression threshold (here we will use 10) for the processed query data.
tooth.epi.E13.5.active.pathway.ranking.metric<-get_pathway_ranking_metric(active.pathway.path = tooth.epi.E13.5.active.pathway, processed.query.data = tooth.epi.E13.5.processed.data, high.exp.th = 10)
## Plot the ranking metric result (i.e., activity score and number of downstream transcription factors) in a 2D plane
display_pathway_ranking_metric(pathway.ranking.metric = tooth.epi.E13.5.active.pathway.ranking.metric)
## To separate the top ranked pathways we can do this
abline(v=45, h=0.2, lty=2, col="black")

}
\author{
Md Humayun Kabir <humayun.mkabir@gmail.com>
}
