% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tooth.epi.E13.5}
\alias{tooth.epi.E13.5}
\title{Data from mouse embryonic dental epithelium cell at E13.5,
contains transcriptomic profiles microarray data of mouse tooth epithelium cell at embryonic day E13.5.
The cell has 3 biological replicates.}
\format{
An object of matrix containing in normalized and log2 format gene expression value of mouse tooth epithelium cell.
}
\usage{
data(tooth.epi.E13.5)
}
\description{
Data from mouse embryonic dental epithelium cell at E13.5,
contains transcriptomic profiles microarray data of mouse tooth epithelium cell at embryonic day E13.5.
The cell has 3 biological replicates.
}
\examples{
head(tooth.epi.E13.5)
}
\references{
Murphy P et al, Development 2018 Jan 9; 145(1)
(\href{https://www.ncbi.nlm.nih.gov/pubmed/29217756}{PubMed})
}
\keyword{datasets}
